FastApi/OpenWeatherMap/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
requests==2.28.2
python-dotenv==0.21.1
jinja2==3.1.2
pydantic==1.10.6